requirements.txt
appdirs==1.4.3
dj-database-url==0.4.2
dj-static==0.0.6
Django==1.11
django-override-settings==1.2
django-webpack-loader==0.5.0
djangorestframework==3.6.2
gunicorn==19.7.1
nodejs==0.1.1
npm==0.1.1
optional-django==0.1.0
packaging==16.8
pyparsing==2.2.0
pytz==2017.2
six==1.10.0
whitenoise==3.3.0
psycopg2==2.6.2
Pillow==4.2.0 